// portfolio/src/features/frameworks/types.ts
export type FrameworksContent = {
  heroTitle: string
  heroTitleEn?: string
  heroSubtitle: string
  heroSubtitleEn?: string

  highlightsTitle: string
  highlightsTitleEn?: string
  highlightsBody: string
  highlightsBodyEn?: string

  reactTitle: string
  reactTitleEn?: string
  reactBody: string
  reactBodyEn?: string

  apiTitle: string
  apiTitleEn?: string
  apiBody: string
  apiBodyEn?: string

  mobileTitle: string
  mobileTitleEn?: string
  mobileBody: string
  mobileBodyEn?: string

  toolingTitle: string
  toolingTitleEn?: string
  toolingBody: string
  toolingBodyEn?: string

  nowTitle: string
  nowTitleEn?: string
  nowBody: string
  nowBodyEn?: string
}